pyproject.toml
[project]
name = "cloud-nettest-framework"
version = "0.1.0"
description = "Multi-cloud network testing framework (AWS/Azure/GCP probes) targeting OCI Object Storage and more"
authors = [{name = "Matthew Schwen"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.9.0",
  "rich>=13.0.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "httpx>=0.25.0",
  "asyncssh>=2.14.0",
  "dnspython>=2.4.0",
  "cryptography>=41.0.0",
  "pandas>=2.1.0",
  "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "ruff>=0.1.0",
  "mypy>=1.6.0",
]

[project.scripts]
cnf = "cnf.cli:app"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
